+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - add authentication fallback to ssh-key in case the ssh-agent is enabled but failed to authenticate
-- improve ssh log messages to clearly indicate the various authentication strategies and fallbacks
+- improve authentication error message to clearly indicate the various strategies failures
 
 ## [14.0.3] - 2019-03-12